@@ -4,15 +4,9 @@ import 'dart:math';
 import 'package:collection/collection.dart'
     show IterableExtension, ListEquality;
 import 'package:dio/dio.dart';
-import 'package:rate_limiter/rate_limiter.dart';
 import 'package:rxdart/rxdart.dart';
 import 'package:stream_chat/src/client/retry_queue.dart';
-import 'package:stream_chat/src/core/error/error.dart';
-import 'package:stream_chat/src/core/models/attachment_file.dart';
-import 'package:stream_chat/src/core/models/channel_state.dart';
-import 'package:stream_chat/src/core/models/user.dart';
 import 'package:stream_chat/src/core/util/utils.dart';
-import 'package:stream_chat/src/event_type.dart';
 import 'package:stream_chat/stream_chat.dart';
 
 /// Class that manages a specific channel.
@@ -825,7 +819,7 @@ class Channel {
     final now = DateTime.now();
     final user = _client.state.currentUser;
 
-    final latestReactions = [...message.latestReactions ?? <Reaction>[]];
+    var latestReactions = [...message.latestReactions ?? <Reaction>[]];
     if (enforceUnique) {
       latestReactions.removeWhere((it) => it.userId == user!.id);
     }
@@ -839,10 +833,17 @@ class Channel {
       extraData: extraData,
     );
 
-    // Inserting at the 0th index as it's the latest reaction
-    latestReactions.insert(0, newReaction);
-    final ownReactions = [...latestReactions]
-      ..removeWhere((it) => it.userId != user!.id);
+    latestReactions = (latestReactions
+          // Inserting at the 0th index as it's the latest reaction
+          ..insert(0, newReaction))
+        .take(10)
+        .toList();
+    final ownReactions = enforceUnique
+        ? <Reaction>[newReaction]
+        : <Reaction>[
+            ...message.ownReactions ?? [],
+            newReaction,
+          ];
 
     final newMessage = message.copyWith(
       reactionCounts: {...message.reactionCounts ?? <String, int>{}}
@@ -882,7 +883,6 @@ class Channel {
     Reaction reaction,
   ) async {
     final type = reaction.type;
-    final user = _client.state.currentUser;
 
     final reactionCounts = {...message.reactionCounts ?? <String, int>{}};
     if (reactionCounts.containsKey(type)) {
@@ -899,8 +899,11 @@ class Channel {
           r.type == reaction.type &&
           r.messageId == reaction.messageId);
 
-    final ownReactions = [...latestReactions]
-      ..removeWhere((it) => it.userId != user!.id);
+    final ownReactions = message.ownReactions
+      ?..removeWhere((r) =>
+          r.userId == reaction.userId &&
+          r.type == reaction.type &&
+          r.messageId == reaction.messageId);
 
     final newMessage = message.copyWith(
       reactionCounts: reactionCounts..removeWhere((_, value) => value == 0),
@@ -1268,6 +1271,21 @@ class Channel {
         pagination: pagination,
       );
 
+  /// Query channel banned users.
+  Future<QueryBannedUsersResponse> queryBannedUsers({
+    Filter? filter,
+    List<SortOption>? sort,
+    PaginationParams? pagination,
+  }) {
+    _checkInitialized();
+    filter ??= Filter.equal('channel_cid', cid!);
+    return _client.queryBannedUsers(
+      filter: filter,
+      sort: sort,
+      pagination: pagination,
+    );
+  }
+
   /// Mutes the channel.
   Future<EmptyResponse> mute({Duration? expiration}) {
     _checkInitialized();
@@ -1281,7 +1299,15 @@ class Channel {
   }
 
   /// Bans the user with given [userID] from the channel.
+  @Deprecated("Use 'banMember' instead. This method will be removed in v4.0.0")
   Future<EmptyResponse> banUser(
+    String userID,
+    Map<String, dynamic> options,
+  ) =>
+      banMember(userID, options);
+
+  /// Bans the member with given [userID] from the channel.
+  Future<EmptyResponse> banMember(
     String userID,
     Map<String, dynamic> options,
   ) async {
@@ -1295,7 +1321,12 @@ class Channel {
   }
 
   /// Remove the ban for the user with given [userID] in the channel.
-  Future<EmptyResponse> unbanUser(String userID) async {
+  @Deprecated(
+      "Use 'unbanMember' instead. This method will be removed in v4.0.0")
+  Future<EmptyResponse> unbanUser(String userID) => unbanMember(userID);
+
+  /// Remove the ban for the member with given [userID] in the channel.
+  Future<EmptyResponse> unbanMember(String userID) async {
     _checkInitialized();
     return _client.unbanUser(userID, {
       'type': type,
@@ -1466,6 +1497,10 @@ class ChannelClientState {
 
     _listenMemberRemoved();
 
+    _listenMemberBanned();
+
+    _listenMemberUnbanned();
+
     _startCleaning();
 
     _startCleaningPinnedMessages();
@@ -1498,16 +1533,21 @@ class ChannelClientState {
               if (url == null || !url.contains('')) {
                 return false;
               }
-              final uri = Uri.parse(url);
-              if (!uri.host.endsWith('stream-io-cdn.com') ||
-                  uri.queryParameters['Expires'] == null) {
+              try {
+                final uri = Uri.parse(url);
+                if (!uri.host.endsWith('stream-io-cdn.com') ||
+                    uri.queryParameters['Expires'] == null) {
+                  return false;
+                }
+                final secondsFromEpoch =
+                    int.parse(uri.queryParameters['Expires']!);
+                final expiration = DateTime.fromMillisecondsSinceEpoch(
+                  secondsFromEpoch * 1000,
+                );
+                return expiration.isBefore(DateTime.now());
+              } catch (_) {
                 return false;
               }
-              final secondsFromEpoch =
-                  int.parse(uri.queryParameters['Expires']!);
-              final expiration =
-                  DateTime.fromMillisecondsSinceEpoch(secondsFromEpoch * 1000);
-              return expiration.isBefore(DateTime.now());
             }))
         .map((e) => e.id)
         .toList();
@@ -1538,6 +1578,7 @@ class ChannelClientState {
         members: List.from(
           channelState.members..removeWhere((m) => m.userId == user!.id),
         ),
+        read: channelState.read..removeWhere((r) => r.user.id == user!.id),
       ));
     }));
   }
@@ -1561,6 +1602,54 @@ class ChannelClientState {
           ?.deleteMessageByCid(channel.cid);
       truncate();
     }));
+  }
+
+  void _listenMemberBanned() {
+    _subscriptions.add(_channel
+        .on(EventType.userBanned)
+        .where((it) => it.cid != null) // filters channel ban from app ban
+        .listen(
+      (event) async {
+        final user = event.user!;
+        final member = await _channel
+            .queryMembers(filter: Filter.equal('id', user.id))
+            .then((it) => it.members.first);
+
+        _updateMember(member);
+      },
+    ));
+  }
+
+  void _listenMemberUnbanned() {
+    _subscriptions.add(_channel
+        .on(EventType.userUnbanned)
+        .where((it) => it.cid != null) // filters channel ban from app ban
+        .listen(
+      (event) async {
+        final user = event.user!;
+        final member = await _channel
+            .queryMembers(filter: Filter.equal('id', user.id))
+            .then((it) => it.members.first);
+
+        _updateMember(member);
+      },
+    ));
+  }
+
+  void _updateMember(Member member) {
+    final currentMembers = [...members];
+    final memberIndex = currentMembers.indexWhere(
+      (m) => m.userId == member.userId,
+    );
+
+    if (memberIndex == -1) return;
+    currentMembers[memberIndex] = member;
+
+    updateChannelState(
+      channelState.copyWith(
+        members: currentMembers,
+      ),
+    );
   }
 
   /// Flag which indicates if [ChannelClientState] contain latest/recent messages or not.
@@ -1604,10 +1693,19 @@ class ChannelClientState {
 
   void _listenReactionDeleted() {
     _subscriptions.add(_channel.on(EventType.reactionDeleted).listen((event) {
-      final userId = _channel.client.state.currentUser!.id;
+      final oldMessage =
+          messages.firstWhereOrNull((it) => it.id == event.message?.id);
+      final reaction = event.reaction;
+      final ownReactions = oldMessage?.ownReactions
+          ?.whereNot((it) =>
+              it.type == reaction?.type &&
+              it.score == reaction?.score &&
+              it.messageId == reaction?.messageId &&
+              it.userId == reaction?.userId &&
+              it.extraData == reaction?.extraData)
+          .toList(growable: false);
       final message = event.message!.copyWith(
-        ownReactions: [...event.message!.latestReactions!]
-          ..removeWhere((it) => it.userId != userId),
+        ownReactions: ownReactions,
       );
       addMessage(message);
     }));
@@ -1615,10 +1713,10 @@ class ChannelClientState {
 
   void _listenReactions() {
     _subscriptions.add(_channel.on(EventType.reactionNew).listen((event) {
-      final userId = _channel.client.state.currentUser!.id;
+      final oldMessage =
+          messages.firstWhereOrNull((it) => it.id == event.message?.id);
       final message = event.message!.copyWith(
-        ownReactions: [...event.message!.latestReactions!]
-          ..removeWhere((it) => it.userId != userId),
+        ownReactions: oldMessage?.ownReactions,
       );
       addMessage(message);
     }));
@@ -1631,10 +1729,11 @@ class ChannelClientState {
       EventType.reactionUpdated,
     )
         .listen((event) {
-      final userId = _channel.client.state.currentUser!.id;
+      final oldMessage =
+          messages.firstWhereOrNull((it) => it.id == event.message?.id);
+
       final message = event.message!.copyWith(
-        ownReactions: [...event.message!.latestReactions!]
-          ..removeWhere((it) => it.userId != userId),
+        ownReactions: oldMessage?.ownReactions,
       );
       addMessage(message);
 
@@ -1729,7 +1828,13 @@ class ChannelClientState {
       if (replyCount == null || replyCount == 0) return;
 
       addMessage(parentMessage.copyWith(replyCount: replyCount - 1));
-      updateThreadInfo(parentId, threads[parentId]!..remove(message));
+      updateThreadInfo(
+        parentId,
+        threads[parentId]!
+          ..removeWhere(
+            (e) => e.id == message.id,
+          ),
+      );
     } else {
       // Remove regular message
       final allMessages = [...messages];
