@@ -1,5 +1,7 @@
-import 'package:stream_chat/src/core/api/responses.dart';
+import 'dart:convert';
+
 import 'package:stream_chat/src/core/http/stream_http_client.dart';
+import 'package:stream_chat/stream_chat.dart';
 
 /// Defines the api dedicated to moderation operations
 class ModerationApi {
@@ -124,5 +126,25 @@ class ModerationApi {
       },
     );
     return EmptyResponse.fromJson(response.data);
+  }
+
+  /// Queries banned users.
+  Future<QueryBannedUsersResponse> queryBannedUsers({
+    Filter? filter,
+    List<SortOption>? sort,
+    PaginationParams? pagination,
+  }) async {
+    final response = await _client.get(
+      '/query_banned_users',
+      queryParameters: {
+        'payload': jsonEncode({
+          if (sort != null) 'sort': sort,
+          if (filter != null) 'filter_conditions': filter,
+          if (pagination != null) ...pagination.toJson(),
+        }),
+      },
+    );
+
+    return QueryBannedUsersResponse.fromJson(response.data);
   }
 }
