@@ -1,7 +1,4 @@
 import 'package:json_annotation/json_annotation.dart';
-import 'package:stream_chat/src/core/models/device.dart';
-import 'package:stream_chat/src/core/models/mute.dart';
-import 'package:stream_chat/src/core/models/user.dart';
 import 'package:stream_chat/src/core/util/serializer.dart';
 import 'package:stream_chat/stream_chat.dart';
 
@@ -29,6 +26,7 @@ class OwnUser extends User {
     bool online = false,
     Map<String, Object?> extraData = const {},
     bool banned = false,
+    DateTime? banExpires,
     List<String> teams = const [],
     String? language,
   }) : super(
@@ -42,6 +40,7 @@ class OwnUser extends User {
           online: online,
           extraData: extraData,
           banned: banned,
+          banExpires: banExpires,
           teams: teams,
           language: language,
         );
@@ -78,6 +77,7 @@ class OwnUser extends User {
     bool? online,
     Map<String, Object?>? extraData,
     bool? banned,
+    DateTime? banExpires,
     List<String>? teams,
     List<Mute>? channelMutes,
     List<Device>? devices,
@@ -89,11 +89,12 @@ class OwnUser extends User {
       OwnUser(
         id: id ?? this.id,
         role: role ?? this.role,
-        /* if null, it will be retrieved from extraData['name']*/
+        // if null, it will be retrieved from extraData['name']
         name: name,
-        /* if null, it will be retrieved from extraData['image']*/
+        // if null, it will be retrieved from extraData['image']
         image: image,
         banned: banned ?? this.banned,
+        banExpires: banExpires ?? this.banExpires,
         createdAt: createdAt ?? this.createdAt,
         updatedAt: updatedAt ?? this.updatedAt,
         lastActive: lastActive ?? this.lastActive,
@@ -139,7 +140,7 @@ class OwnUser extends User {
   @JsonKey(includeIfNull: false)
   final List<Mute> mutes;
 
-  /// List of users muted by the user.
+  /// List of channels muted by the user.
   @JsonKey(includeIfNull: false)
   final List<Mute> channelMutes;
 
