@@ -1,20 +1,8 @@
 import 'package:mocktail/mocktail.dart';
-import 'package:stream_chat/src/client/client.dart';
 import 'package:stream_chat/src/core/api/device_api.dart';
-import 'package:stream_chat/src/core/api/requests.dart';
-import 'package:stream_chat/src/core/api/responses.dart';
-import 'package:stream_chat/src/core/error/error.dart';
 import 'package:stream_chat/src/core/http/token.dart';
-import 'package:stream_chat/src/core/models/channel_model.dart';
-import 'package:stream_chat/src/core/models/event.dart';
-import 'package:stream_chat/src/core/models/filter.dart';
-import 'package:stream_chat/src/core/models/message.dart';
-import 'package:stream_chat/src/core/models/own_user.dart';
-import 'package:stream_chat/src/core/models/user.dart';
-import 'package:stream_chat/src/event_type.dart';
-import 'package:stream_chat/src/ws/connection_status.dart';
+import 'package:stream_chat/src/core/models/banned_user.dart';
 import 'package:stream_chat/stream_chat.dart';
-import 'package:test/scaffolding.dart';
 import 'package:test/test.dart';
 
 import '../fakes.dart';
@@ -993,6 +981,36 @@ void main() {
             pagination: any(named: 'pagination'),
           )).called(1);
       verifyNoMoreInteractions(api.user);
+    });
+
+    test('`.queryBannedUsers`', () async {
+      final bans = List.generate(
+        3,
+        (index) => BannedUser(
+          user: User(id: 'test-user-id-$index'),
+          bannedBy: User(id: 'test-user-id-${index + 1}'),
+        ),
+      );
+
+      const cid = 'message:nice-channel';
+      final filter = Filter.equal('channel_cid', cid);
+
+      when(() => api.moderation.queryBannedUsers(
+            filter: filter,
+            sort: any(named: 'sort'),
+            pagination: any(named: 'pagination'),
+          )).thenAnswer((_) async => QueryBannedUsersResponse()..bans = bans);
+
+      final res = await client.queryBannedUsers(filter: filter);
+      expect(res, isNotNull);
+      expect(res.bans.length, bans.length);
+
+      verify(() => api.moderation.queryBannedUsers(
+            filter: filter,
+            sort: any(named: 'sort'),
+            pagination: any(named: 'pagination'),
+          )).called(1);
+      verifyNoMoreInteractions(api.moderation);
     });
 
     test('`.search`', () async {
@@ -2312,6 +2330,33 @@ void main() {
             set: {'pinned': false},
           )).called(1);
       verifyNoMoreInteractions(api.message);
+    });
+
+    test('`.enrichUrl`', () async {
+      const url =
+          'https://www.techyourchance.com/finite-state-machine-with-unit-tests-real-world-example';
+
+      when(() => api.general.enrichUrl(url)).thenAnswer(
+        (_) async => OGAttachmentResponse()
+          ..type = 'image'
+          ..ogScrapeUrl = url
+          ..authorName = 'TechYourChance'
+          ..title = 'Finite State Machine with Unit Tests: Real World Example',
+      );
+
+      final res = await client.enrichUrl(url);
+
+      expect(res, isNotNull);
+      expect(res.type, 'image');
+      expect(res.ogScrapeUrl, url);
+      expect(res.authorName, 'TechYourChance');
+      expect(
+        res.title,
+        'Finite State Machine with Unit Tests: Real World Example',
+      );
+
+      verify(() => api.general.enrichUrl(url)).called(1);
+      verifyNoMoreInteractions(api.general);
     });
 
     test(
