@@ -2,11 +2,8 @@ import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
 import 'package:shimmer/shimmer.dart';
 import 'package:stream_chat_flutter/src/attachment/attachment_title.dart';
-import 'package:stream_chat_flutter/src/attachment/attachment_upload_state_builder.dart';
 import 'package:stream_chat_flutter/src/attachment/attachment_widget.dart';
-import 'package:stream_chat_flutter/src/theme/themes.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
-import 'package:stream_chat_flutter_core/stream_chat_flutter_core.dart';
 
 /// Widget for showing an image attachment
 class ImageAttachment extends AttachmentWidget {
@@ -141,7 +138,9 @@ class ImageAttachment extends AttachmentWidget {
                                 return StreamChannel(
                                   channel: channel,
                                   child: FullScreenMedia(
-                                    mediaAttachments: [attachment],
+                                    mediaAttachments: message.attachments,
+                                    startIndex:
+                                        message.attachments.indexOf(attachment),
                                     userName: message.user?.name,
                                     message: message,
                                     onShowMessage: onShowMessage,
