@@ -1,8 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:stream_chat_flutter/src/attachment/attachment_title.dart';
 import 'package:stream_chat_flutter/src/attachment/attachment_widget.dart';
-import 'package:stream_chat_flutter/src/full_screen_media.dart';
-import 'package:stream_chat_flutter/src/theme/themes.dart';
 import 'package:stream_chat_flutter/src/video_thumbnail_image.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
 
@@ -87,7 +85,9 @@ class VideoAttachment extends AttachmentWidget {
                           builder: (_) => StreamChannel(
                             channel: channel,
                             child: FullScreenMedia(
-                              mediaAttachments: [attachment],
+                              mediaAttachments: message.attachments,
+                              startIndex:
+                                  message.attachments.indexOf(attachment),
                               userName: message.user?.name,
                               message: message,
                               onShowMessage: onShowMessage,
