@@ -245,7 +245,7 @@ class _MessageViewState extends State<MessageView> {
                 ),
               ],
             ),
-          )
+          ),
         ],
       );
 }
